test(epub): extract scholarly reference family
